+++ crates/nostr/src/event/unsigned.rs
@@ -107,7 +107,7 @@ impl UnsignedEvent {
     #[inline]
     #[cfg(feature = "std")]
     pub fn sign_with_keys(self, keys: &Keys) -> Result<Event, Error> {
-        self.sign_with_ctx(&SECP256K1, &mut OsRng, keys)
+        self.sign_with_ctx(SECP256K1, &mut OsRng, keys)
     }
 
     /// Sign an unsigned event with [`Keys`] signer
@@ -136,7 +136,7 @@ impl UnsignedEvent {
     #[inline]
     #[cfg(feature = "std")]
     pub fn add_signature(self, sig: Signature) -> Result<Event, Error> {
-        self.add_signature_with_ctx(&SECP256K1, sig)
+        self.add_signature_with_ctx(SECP256K1, sig)
     }
 
     /// Add signature to unsigned event