+++ crates/nostr/src/util/mod.rs
@@ -8,12 +8,8 @@ use alloc::string::String;
 use core::fmt::Debug;
 
 #[cfg(feature = "std")]
-use once_cell::sync::Lazy; // TODO: use `std::sync::LazyLock` when MSRV >= 1.80.0
-#[cfg(feature = "std")]
-use secp256k1::rand::rngs::OsRng;
+use secp256k1::global::GlobalContext;
 use secp256k1::{ecdh, Parity, PublicKey as NormalizedPublicKey, XOnlyPublicKey};
-#[cfg(feature = "std")]
-use secp256k1::{All, Secp256k1};
 use serde::de::DeserializeOwned;
 use serde::Serialize;
 
@@ -43,11 +39,7 @@ pub fn generate_shared_key(
 
 /// Secp256k1 global context
 #[cfg(feature = "std")]
-pub static SECP256K1: Lazy<Secp256k1<All>> = Lazy::new(|| {
-    let mut ctx = Secp256k1::new();
-    ctx.randomize(&mut OsRng);
-    ctx
-});
+pub static SECP256K1: &GlobalContext = secp256k1::global::SECP256K1;
 
 /// JSON util
 pub trait JsonUtil: Sized + Serialize + DeserializeOwned